requirement.txt
streamlit
google.generativeai
python-dotenv
matplotlib-venn
PyPDF2
Pillow